@@ -1,49 +1,42 @@
 [project]
 name = "marvel-characters"
 description = "Marvel characters project"
-requires-python = ">=3.11, <3.12"
+requires-python = ">=3.12, <3.13"
 dependencies = [
     "mlflow==3.1.1",
     "cffi==1.17.1",
-    "cloudpickle==3.1.0",
-    "matplotlib==3.9.2",
+    "cloudpickle==3.1.1",
     "numpy==1.26.4",
-    "pandas==2.2.3",
-    "psutil==6.0.0",
-    "pyarrow==14.0.1",
-    "scikit-learn==1.5.2",
-    "lightgbm==4.5.0",
-    "scipy==1.14.1",
-    "databricks-sdk==0.32.0",
-    "pydantic==2.9.2",
+    "pandas==2.3.0",
+    "pyarrow==17.0.0",
+    "databricks-feature-engineering==0.12.1",
+    "scikit-learn==1.7.0",
+    "lightgbm==4.6.0",
+    "scipy==1.16.0",
+    "databricks-sdk==0.55.0",
+    "pydantic==2.11.7",
     "loguru==0.7.3",
-    "python-dotenv>=1.1.0",
+    "python-dotenv==1.1.1",
+    "ray[tune]==2.47.1",
+    "optuna>=4.4.0",
+    "sdv>=1.24.0",
 ]
 
 dynamic = ['version']
 
 [project.optional-dependencies]
-dev = [
-    "databricks-connect==15.4.5",
-    "pyspark==3.5.5",
+dev = ["databricks-connect>=16.0, <17",
        "ipykernel>=6.29.5, <7",
-       "pip>=24.2",
-       "pre-commit>=4.1.0, <5",
-       "pytest>=8.3.4, <9",
-    ]
+       "pip>=25.0.1, <26",
+       "pre-commit>=4.1.0, <5"]
 
-test = [
-    "pre-commit>=4.1.0, <5",
-    "delta-spark==3.3.0",
-    "pyspark==3.5.5",
-    "pytest>=8.3.5",
-    "pytest-cov >=6.1.0",
-    "pydantic-settings>=2.8.1",
-]
+ci = ["pre-commit>=4.1.0, <5"]
 
 [build-system]
-requires = ["setuptools>=72.0"]
+requires = [
+  "setuptools>=72.0"]
 build-backend = "setuptools.build_meta"
+
 
 [tool.setuptools.packages.find]
 where =["src"]
