@@ -25,7 +25,14 @@ def create_or_refresh_monitoring(config: ProjectConfig, spark: SparkSession, wor
     :param spark: Spark session used for executing SQL queries and transformations.
     :param workspace: Workspace object used for managing quality monitors.
     """
-    inf_table = spark.sql(f"SELECT * FROM {config.catalog_name}.{config.schema_name}.`model-serving-fe_payload`")
+    # Check if custom_model_payload table exists and has data
+    inf_table = spark.sql(f"SELECT * FROM {config.catalog_name}.{config.schema_name}.`custom_model_payload`")
+    inf_count = inf_table.count()
+    logger.info(f"Found {inf_count} records in custom_model_payload table")
+
+    if inf_count == 0:
+        logger.warning("No records found in custom_model_payload table. Monitoring table will be empty.")
+        return
 
     request_schema = StructType(
         [
@@ -91,30 +98,66 @@ def create_or_refresh_monitoring(config: ProjectConfig, spark: SparkSession, wor
     )
 
     test_set = spark.table(f"{config.catalog_name}.{config.schema_name}.test_set")
-    inference_set_skewed = spark.table(f"{config.catalog_name}.{config.schema_name}.inference_data_skewed")
+
+    # Log counts at each step to diagnose where data might be getting filtered out
+    logger.info(f"Records in df_final before join: {df_final.count()}")
+
+    # Check if test_set has data
+    test_count = test_set.count()
+    logger.info(f"Records in test_set: {test_count}")
+
+    # Join with test set
+    df_joined = df_final.join(test_set.select("Id", "Alive"), on="Id", how="left")
+    logger.info(f"Records after join: {df_joined.count()}")
+
+    # Count nulls in key columns
+    null_alive = df_joined.filter(F.col("Alive").isNull()).count()
+    null_prediction = df_joined.filter(F.col("prediction").isNull()).count()
+    logger.info(f"Records with null Alive: {null_alive}, Records with null prediction: {null_prediction}")
 
     df_final_with_status = (
-        df_final.join(test_set.select("Id", "Alive"), on="Id", how="left")
-        .withColumnRenamed("Alive", "alive_test")
-        .join(inference_set_skewed.select("Id", "Alive"), on="Id", how="left")
-        .withColumnRenamed("Alive", "alive_inference")
-        .select("*", F.coalesce(F.col("alive_test"), F.col("alive_inference")).alias("alive"))
-        .drop("alive_test", "alive_inference")
+        df_joined.withColumnRenamed("Alive", "alive")
         .withColumn("alive", F.col("alive").cast("int"))
         .withColumn("prediction", F.col("prediction").cast("int"))
-        .dropna(subset=["alive", "prediction"])
     )
 
-    marvel_features = spark.table(f"{config.catalog_name}.{config.schema_name}.marvel_features")
+    # Make dropna optional if we're losing all data
+    df_with_valid_values = df_final_with_status.dropna(subset=["alive", "prediction"])
+    valid_count = df_with_valid_values.count()
+    logger.info(f"Records with valid alive and prediction values: {valid_count}")
 
-    df_final_with_features = df_final_with_status.join(marvel_features, on="Id", how="left")
+    # If we lost all data after dropna, use the data before dropna
+    if valid_count > 0:
+        df_final_with_status = df_with_valid_values
+        logger.info("Using records with valid alive and prediction values")
+    else:
+        logger.warning("All records have null alive or prediction values. Using records with potential nulls.")
 
-    df_final_with_features = df_final_with_features.withColumn("Height", F.col("Height").cast("double"))
-    df_final_with_features = df_final_with_features.withColumn("Weight", F.col("Weight").cast("double"))
+    # Check if we have data to proceed
+    status_count = df_final_with_status.count()
+    logger.info(f"Records after processing status: {status_count}")
 
-    df_final_with_features.write.format("delta").mode("append").saveAsTable(
+    if status_count == 0:
+        logger.warning("No records to write to monitoring table after processing.")
+        return
+
+    # We'll skip the marvel_features join since we already have Height and Weight in df_final
+    # Just ensure Height and Weight are properly cast to double
+    df_final_with_status = df_final_with_status.withColumn("Height", F.col("Height").cast("double"))
+    df_final_with_status = df_final_with_status.withColumn("Weight", F.col("Weight").cast("double"))
+
+    # Log the final count before writing
+    final_count = df_final_with_status.count()
+    logger.info(f"Final record count before writing to monitoring table: {final_count}")
+
+    # Write to the monitoring table
+    df_final_with_status.write.format("delta").mode("append").saveAsTable(
         f"{config.catalog_name}.{config.schema_name}.model_monitoring"
     )
+
+    # Verify data was written
+    written_count = spark.table(f"{config.catalog_name}.{config.schema_name}.model_monitoring").count()
+    logger.info(f"Records in monitoring table after write: {written_count}")
 
     try:
         workspace.quality_monitors.get(f"{config.catalog_name}.{config.schema_name}.model_monitoring")
@@ -157,3 +200,5 @@ def create_monitoring_table(config: ProjectConfig, spark: SparkSession, workspac
 
     # Important to update monitoring
     spark.sql(f"ALTER TABLE {monitoring_table} SET TBLPROPERTIES (delta.enableChangeDataFeed = true);")
+
+    logger.info("Lakehouse monitoring table is created.")
