@@ -1,55 +1,5 @@
 # Databricks notebook source
-# !pip install house_price-1.1.1-py3-none-any.whl
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC ## Feature Importance
-
-# COMMAND ----------
-
-import pandas as pd
-from pyspark.sql.functions import col
-from pyspark.sql.functions import current_timestamp, to_utc_timestamp
-import numpy as np
-from pyspark.sql import SparkSession
-from loguru import logger
-from marvel_characters.config import ProjectConfig
-
-spark = SparkSession.builder.getOrCreate()
-
-# Load configuration
-config = ProjectConfig.from_yaml(config_path="../project_config_marvel.yml", env="dev")
-spark = SparkSession.builder.getOrCreate()
-
-train_set = spark.table(f"{config.catalog_name}.{config.schema_name}.train_set").toPandas()
-test_set = spark.table(f"{config.catalog_name}.{config.schema_name}.test_set").toPandas()
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC ## Generate Synthetic Data
-
-# COMMAND ----------
-from marvel_characters.data_processor import generate_synthetic_data
-
-inference_data_skewed = generate_synthetic_data(train_set, drift= True, num_rows=100)
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC ### Create Tables and Update marvel_features_online
-
-# COMMAND ----------
-
-inference_data_skewed_spark = spark.createDataFrame(inference_data_skewed).withColumn(
-    "update_timestamp_utc", to_utc_timestamp(current_timestamp(), "UTC")
-)
-
-inference_data_skewed_spark.write.mode("overwrite").saveAsTable(
-    f"{config.catalog_name}.{config.schema_name}.inference_data_skewed"
-)
-
+# %pip install marvel_characters-0.1.0-py3-none-any.whl
 # COMMAND ----------
 
 # MAGIC %md
@@ -57,13 +7,13 @@ inference_data_skewed_spark.write.mode("overwrite").saveAsTable(
 
 # COMMAND ----------
 
-import pandas as pd
 from pyspark.sql.functions import col
-from pyspark.sql.functions import current_timestamp, to_utc_timestamp
-import numpy as np
 from pyspark.sql import SparkSession
+import os
+from pyspark.dbutils import DBUtils
 
 from marvel_characters.config import ProjectConfig
+from marvel_characters.utils import is_databricks
 
 spark = SparkSession.builder.getOrCreate()
 
@@ -74,19 +24,20 @@ test_set = spark.table(f"{config.catalog_name}.{config.schema_name}.test_set") \
                         .withColumn("Id", col("Id").cast("string")) \
                         .toPandas()
 
-
-inference_data_skewed = spark.table(f"{config.catalog_name}.{config.schema_name}.inference_data_skewed") \
-                        .withColumn("Id", col("Id").cast("string")) \
-                        .toPandas()
-
-
 # COMMAND ----------
-
-token = dbutils.notebook.entry_point.getDbutils().notebook().getContext().apiToken().get()
-host = spark.conf.get("spark.databricks.workspaceUrl")
-
+if is_databricks():
+    from pyspark.dbutils import DBUtils
+    dbutils = DBUtils(spark)
+    os.environ["DBR_TOKEN"] = dbutils.notebook.entry_point.getDbutils().notebook().getContext().apiToken().get()
+    os.environ["DBR_HOST"] = spark.conf.get("spark.databricks.workspaceUrl")
+else:
+    from dotenv import load_dotenv
+    load_dotenv()
+    # DBR_TOKEN and DBR_HOST should be set in your .env file
+    assert os.environ.get("DBR_TOKEN"), "DBR_TOKEN must be set in your environment or .env file."
+    assert os.environ.get("DBR_HOST"), "DBR_HOST must be set in your environment or .env file."
+    profile = os.environ["PROFILE"]
 # COMMAND ----------
-
 
 from databricks.sdk import WorkspaceClient
 import requests
@@ -113,11 +64,11 @@ def send_request_https(dataframe_record):
     """
     Sends a request to the model serving endpoint using HTTPS.
     """
-    serving_endpoint = f"https://{host}/serving-endpoints/marvel-characters-model-serving/invocations"
+    serving_endpoint = f"https://{os.environ['DBR_HOST']}/serving-endpoints/marvel-characters-model-serving/invocations"
     
     response = requests.post(
         serving_endpoint,
-        headers={"Authorization": f"Bearer {token}"},
+        headers={"Authorization": f"Bearer {os.environ['DBR_TOKEN']}"},
         json={"dataframe_records": dataframe_record},
     )
     return response.status_code, response.text
@@ -126,7 +77,7 @@ def send_request_workspace(dataframe_record):
     """
     Sends a request to the model serving endpoint using workspace client.
     """
-    serving_endpoint = f"https://{host}/serving-endpoints/marvel-characters-model-serving/invocations"
+    serving_endpoint = f"https://{os.environ["DBR_HOST"]}/serving-endpoints/marvel-characters-model-serving/invocations"
     
     response = workspace.serving_endpoints.query(
         name="marvel-characters-model-serving",
@@ -147,4 +98,21 @@ for i in range(len(dataframe_records)):
     status_code, response_text = send_request_https(dataframe_records[i])
     print(f"Response Status: {status_code}")
     print(f"Response Text: {response_text}")
-    time.sleep(0.2) 
+    time.sleep(0.2)
+    
+
+# COMMAND ----------
+
+from databricks.connect import DatabricksSession
+from databricks.sdk import WorkspaceClient
+
+from marvel_characters.config import ProjectConfig
+from marvel_characters.monitoring import create_or_refresh_monitoring
+
+spark = DatabricksSession.builder.getOrCreate()
+workspace = WorkspaceClient()
+
+# Load configuration
+config = ProjectConfig.from_yaml(config_path="../project_config_marvel.yml", env="dev")
+
+create_or_refresh_monitoring(config=config, spark=spark, workspace=workspace)
